Move and rename Legende component
