+++ components/simulation-menu/legende/Legende.module.scss
