+++ components/simulation-menu/legende/index.js
@@ -1,0 +1,1 @@
+export { default as Legende } from "./Legende";
